Тест по Фибоначи

+++ 03.AlgebraicAlgorithm/src/main/power/PowerAdvance.java
@@ -2,12 +2,12 @@ package main.power;
 
 public class PowerAdvance {
 
-    public long power(int a, int n) {
-        long result = 1;
+    public int power(int a, int n) {
+        int result = 1;
         long temp = a;
-        for (n/=2; n >= 1; n/=2) {
+        for (n /= 2; n >= 1; n /= 2) {
             temp *= temp;
-            if(n%2 == 1) {
+            if (n % 2 == 1) {
                 result *= temp;
             }
         }
