+++ 03.AlgebraicAlgorithm/src/main/power/PowerWithMultiplication.java
@@ -1,14 +1,15 @@
 package main.power;
 
 public class PowerWithMultiplication {
-    public  long power(int a, int n) {
+
+    public int power(int a, int n) {
         int p = 1;
-        long result = a;
-        for (; p <= n/2; p *= 2) {
+        int result = a;
+        for (; p <= n / 2; p *= 2) {
             result *= result;
         }
         for (int i = p - n; i > 0; i--) {
-           result *= a;
+            result *= a;
         }
         return result;
     }
